Refactor each client into their own module to prepare for localllm
