+++ tinytroupe/config.ini
@@ -14,7 +14,7 @@ AZURE_API_VERSION=2023-05-15
 # Model parameters
 #
 
-MODEL=gpt-4o
+MODEL=gpt-4o-mini
 MAX_TOKENS=4000
 TEMPERATURE=0.3
 FREQ_PENALTY=0.0
