+++ tinytroupe/factory.py
@@ -5,7 +5,7 @@ import logging
 import copy
 logger = logging.getLogger("tinytroupe")
 
-from tinytroupe import openai_utils
+from tinytroupe.clients import client
 from tinytroupe.agent import TinyPerson
 import tinytroupe.utils as utils
 from tinytroupe.control import transactional
@@ -131,7 +131,7 @@ class TinyPersonFactory(TinyFactory):
 
         messages.append({"role": "user", "content": user_prompt})
 
-        response = openai_utils.client().send_message(messages)
+        response = client().send_message(messages)
 
         if response is not None:
             result = utils.extract_json(response["content"])
@@ -215,7 +215,7 @@ class TinyPersonFactory(TinyFactory):
         due too a technicality - otherwise, the agent creation would be skipped during cache reutilization, and
         we don't want that.
         """
-        return openai_utils.client().send_message(messages, temperature=temperature)
+        return client().send_message(messages, temperature=temperature)
     
     @transactional
     def _setup_agent(self, agent, configuration):
